refactor: 边框2 tsx改造
