+++ packages/docs/docs/Border/BorderBox2/demo.vue
@@ -7,3 +7,7 @@
     </dv-border-box2>
   </div>
 </template>
+
+<script lang="ts" setup>
+import { BorderBox2 as DvBorderBox2 } from '@kjgl77/datav-vue3'
+</script>
